Removed SKAbstractDrawable Added SKPictureRecorder.EndRecordingAsDrawable()
