+++ binding/Binding/SKPictureRecorder.cs
@@ -37,6 +37,11 @@ namespace SkiaSharp
 			return GetObject<SKPicture> (SkiaApi.sk_picture_recorder_end_recording (Handle));
 		}
 
+		public SKDrawable EndRecordingAsDrawable ()
+		{
+			return GetObject<SKDrawable> (SkiaApi.sk_picture_recorder_end_recording_as_drawable (Handle));
+		}
+
 		public SKCanvas RecordingCanvas => GetObject<SKCanvas> (SkiaApi.sk_picture_get_recording_canvas (Handle), false);
 	}
 }
